+++ builder/osc/common/step_create_tags.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/antihax/optional"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	retry "github.com/hashicorp/packer/common"
+	"github.com/hashicorp/packer/builder/osc/common/retry"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer-plugin-sdk/template/interpolate"
