+++ builder/osc/common/step_run_source_vm.go
@@ -12,7 +12,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/outscale/osc-sdk-go/osc"
 
-	retry "github.com/hashicorp/packer/common"
+	"github.com/hashicorp/packer/builder/osc/common/retry"
 	"github.com/hashicorp/packer/helper/communicator"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
